Improve performance of arbitrary-precision integer parser.
The arbitrary-precision integer parser used to fold digits
onto an Integer one digit at a time. This commit makes it
instead fold as many digits as possible into a machine
register before folding this result onto the Integer
accumulator. This results in an 8x performance improvement
on the 64-bit platform that the author is benchmarking on.
Additionally, this commit fixes some correctness issues with
decUnsignedInt.